+++ CHANGELOG.md
@@ -17,8 +17,12 @@ Upcoming feature release, expected around 1 November 2025.
 ### Added
 
  - #209: Added `novas_gmst()` and `novas_gast()` functions to calculate Greenwich Mean and Apparent (respectively)
-   Sidereal times for a given UT1 date. The new functions are meant to replace the old NOVAS C `sidereal_time()` with
+   Sidereal Times for a given UT1 date. The new functions are meant to replace the old NOVAS C `sidereal_time()` with
    a simpler and more intuitive interface.
+   
+ - #215: New functions (e.g. `novas_diurnal_eop()`) to calculate corrections to the Earth orientation parameters 
+   published by IERS, to include the effect of libration and ocean tides. Such corrections are necessary to include 
+   if needing or using ITRS / TIRS coordinates with accuracy below the milli-arcsecond (mas) level.
 
 ### Changed